Database one-to-many architecture realized. AddCategoryFragment, IconsFragment added

+++ app/src/main/java/ru/maxstelmakh/mymoney/data/models/CategoryModelData.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class CategoryModelData(
     @PrimaryKey(autoGenerate = true)
-    var categoryId: Long = 0,
+    var categoryId: Int = 0,
     var category: String,
     var color: String,
     var image: Int
