+++ app/src/main/java/ru/maxstelmakh/mymoney/domain/model/CategoryModelDomain.kt
@@ -1,7 +1,7 @@
 package ru.maxstelmakh.mymoney.domain.model
 
 data class CategoryModelDomain(
-    var categoryId: Long = 0,
+    var categoryId: Int = 0,
     var category: String,
     var color: String,
     var image: Int
